requirements.txt
aiohttp
python-telegram-bot==21.6
openai
aiosqlite
python-dotenv
